Fix acorn types
