+++ src/utils/uniqueIds.ts
@@ -1,4 +1,4 @@
-import { simple } from 'acorn-walk/dist/walk'
+import { simple } from '../utils/walkers'
 import * as es from 'estree'
 import { NativeStorage } from '../types'
